Relocate checkIsValidZoningCode to utils
